requirements.txt
Flask==2.2.5
Werkzeug==3.0.0
python-dateutil==2.8.2